+++ background.js
@@ -94,6 +94,13 @@ async function onPortMessage(msg, port) {
     case 'compositionstart':
       play('window.compositionstart');
       break;
+
+    case 'enter-fullscreen':
+      play('doc.fullscreenEnter');
+      break;
+    case 'leave-fullscreen':
+      play('doc.fullscreenLeave');
+      break;
     }
     break;
   case 'ready':